+++ bindings/src/account_history/msg/mod.rs
@@ -30,9 +30,9 @@ pub mod query_resp {
     pub use user_value_response::UserValueResponse;
 
     mod staked_assets_response;
-    pub use staked_assets_response::StakedAssetsResponse;
     pub use staked_assets_response::StakeAssetBalanceBreakdown;
-    
+    pub use staked_assets_response::StakedAssetsResponse;
+
     pub mod earn {
         mod get_eden_boost_earn_details_resp;
         pub use get_eden_boost_earn_details_resp::GetEdenBoostEarnProgramResp;
@@ -53,6 +53,6 @@ pub mod query_resp {
         mod get_masterchef_pending_rewards;
         mod get_masterchef_pool_apr_response;
         pub use get_masterchef_pending_rewards::GetMasterchefUserPendingRewardResponse;
-        pub use get_masterchef_pool_apr_response::QueryPoolAprsResponse;
+        pub use get_masterchef_pool_apr_response::{MasterChefPoolAprResponse, PoolAprValue};
     }
 }
